backend/requirements.txt
fastapi
uvicorn
PyPDF2
python-docx
python-multipart